H2M1/FlaskApp_Template/requirements.txt
flask
python-dotenv
ipykernel
torch
pillow
transformers
openai